Added the HTML generation
App is now able to generate an HTML file with the results. For now, it has a pie chart that indicates the percentage of positive, negative and neutral comments. Besides that, it presents the number of comments and likes, as well as the photo from the post and its description.